requirements-dev.txt
alabaster==0.7.12
allennlp==0.9.0
apipkg==1.5
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.4
blis==0.4.1
boto3==1.12.39
botocore==1.15.39
catalogue==1.0.0
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
commonmark==0.9.1
conllu==1.3.1
coverage==5.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
docutils==0.15.2
editdistance==0.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz#egg=en_core_web_sm
entrypoints==0.3
execnet==1.7.1
Faker==4.0.3
filelock==3.0.12
flaky==3.6.1
Flask==1.1.2
Flask-Cors==3.0.8
ftfy==5.7
fuzzysearch==0.7.1
fuzzywuzzy==0.18.0
gevent==1.5.0
greenlet==0.4.15
h5py==2.10.0
identify==1.4.14
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonnet==0.15.0
jsonpickle==1.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
Morfessor==2.0.6
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.5
nodeenv==1.3.5
notebook==6.0.3
numpy==1.18.2
numpydoc==0.9.2
overrides==2.8.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plac==0.9.6
pluggy==0.13.1
polyglot==16.7.4
pre-commit==2.1.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pycld2==0.41
Pygments==2.6.1
PyICU==2.4.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.3.5
pytest-cov==2.8.1
pytest-forked==1.1.3
pytest-mock==3.1.0
pytest-xdist==1.32.0
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
recommonmark==0.6.0
regex==2020.4.4
requests==2.23.0
responses==0.10.12
s3transfer==0.3.3
sacremoses==0.0.41
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentencepiece==0.1.85
six==1.14.0
snowballstemmer==2.0.0
spacy==2.2.4
Sphinx==3.0.2
sphinx-markdown-tables==0.0.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
srsly==1.0.2
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==7.4.0
tokenizers==0.0.11
toml==0.10.0
torch==1.4.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
transformers==2.4.1
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.18
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
word2number==1.1
wordcloud==1.6.0
zipp==3.1.0
